Gaithersburg qualifiers fixes
Set starting desired position to the current encoder positions at the start of teleop for the shoulder and linear slide so the controller doesn't move.

Adjust the timing of lowering the shoulder for specimen auto because the specimen was not clipping onto the rung.

Add a soft limit to the climber because the hooks could trap the rung against the linear slide motor.