Fixes autonomous specimen and park
The robotController.update method was not being called in every state so the should would keep moving after switching to the drive state. This moves the robotController.update to the end of the state machine loops so it is called every time. Then, each state sets the shoulder and claw commands that need to be sent for that state. If additional controls need to be sent (e.g. linear slide), then those variables will also need to be added and set in every state.