Fixes odometry setup and rotation
The offsets for the odometry pod were incorrect and that was causing the rotation to be centered around the wrong point on (actually outside) the robot. There was also a bug in the DriveToPoint code that didn't take into angle delta greater than pi or less than -pi. If the desired angle difference is more than pi or less than -pi from the current robot angle, then 2 pi is subtracted or added, respectively, to the difference so the shortest path is always taken.

Also, updated the targets in the high basket auto now that the rotation is correct.